Rename complexTransactions file to apiHelpers
